+++ app/drawer.ts
@@ -13,15 +13,11 @@ import {
   verticalBlocks,
   verticalDivisions,
   verticalRatio,
+  imgPath,
 } from './constants';
 import { pathDefinition } from './pathDefinition';
-import * as structureDef from './tools/structure';
-import * as amenitiesDef from './tools/amenities';
-import * as constructionDef from './tools/construction';
-import * as treeDef from './tools/tree';
-import * as flowerDef from './tools/flower';
 import { sweepPath } from './helpers/sweepPath';
-import { createMenu } from './createMenu';
+import { createMenu } from './ui/createMenu';
 import {
   state,
   clearMap,
@@ -34,39 +30,37 @@ import {
   canUndo,
   canRedo,
   objectPositionCommand,
-  drawCommand,
 } from './state';
 import { emitter } from './emitter';
-import { updateBrush, setBrushLineForce, initBrush } from './brush';
+import {
+  updateBrush,
+  setBrushLineForce,
+  initBrush,
+  updatePaintColor,
+  incrementBrush,
+  decrementBrush,
+  cycleBrushHead,
+} from './brush';
+import { modalLayer } from './ui/modal';
+import { showHelpMenu } from './ui/help';
+import { toolState } from './tools/state';
+import { toolCategoryDefinition, init } from './tools';
+import { createButton } from './ui/createButton';
+import { addToLeftToolMenu, createLeftMenu } from './ui/leftMenu';
+import { showMainMenu } from './ui/mainMenu';
+import { getDistanceFromWholeNumber } from './helpers/getDistanceFromWholeNumber';
+import { pointApproximates } from './helpers/pointApproximates';
+import { initLayers, layers } from './layers';
 
 const editor = {
   autosaveMap: null,
   clearAutosave: null,
 };
+
 export function drawer() {
   const backgroundLayer = paper.project.activeLayer;
-  const mapLayer = new paper.Layer();
-  const mapIconLayer = new paper.Layer();
-  const mapOverlayLayer = new paper.Layer();
-  const uiLayer = new paper.Layer();
-  const fixedLayer = new paper.Layer();
-  const cloudLayer = new paper.Layer();
-  const modalLayer = new paper.Layer();
+  initLayers();
   backgroundLayer.applyMatrix = false;
-  cloudLayer.applyMatrix = false;
-  fixedLayer.applyMatrix = false;
-  modalLayer.applyMatrix = false;
-
-  mapLayer.applyMatrix = false;
-  mapLayer.pivot = new paper.Point(0, 0);
-  mapIconLayer.applyMatrix = false;
-  mapIconLayer.pivot = new paper.Point(0, 0);
-  mapOverlayLayer.applyMatrix = false;
-  mapOverlayLayer.pivot = new paper.Point(0, 0);
-
-  // load assets
-  const imgPath = 'static/img/';
-  const toolPrefix = 'tool-';
 
   let brushSizeUI;
   function showBrushSizeUI(isShown) {
@@ -326,13 +320,13 @@ export function drawer() {
   }
 
   function createObjectPreview(objectDefinition, itemData) {
-    mapOverlayLayer.activate();
+    layers.mapOverlayLayer.activate();
     const group = createObjectBase(objectDefinition, itemData);
     return group;
   }
 
   function createObject(objectDefinition, itemData) {
-    mapIconLayer.activate();
+    layers.mapIconLayer.activate();
 
     const group = createObjectBase(objectDefinition, itemData);
     if (objectDefinition.extraObject) {
@@ -403,14 +397,14 @@ export function drawer() {
       // }
 
       this.elements.bound.strokeColor.alpha = 1;
-      const coordinate = mapOverlayLayer.globalToLocal(event.point);
+      const coordinate = layers.mapOverlayLayer.globalToLocal(event.point);
       this.data.prevPosition = this.position;
       this.data.wasMoved = false;
       this.data.clickPivot = coordinate - this.pivot;
       grabObject(coordinate, this);
     };
     group.onMouseDrag = function (event) {
-      const coordinate = mapOverlayLayer.globalToLocal(event.point);
+      const coordinate = layers.mapOverlayLayer.globalToLocal(event.point);
       this.position = (coordinate - this.data.clickPivot).round();
       if (this.position.getDistance(this.data.prevPosition, true) > 0.1) {
         this.data.wasMoved = true;
@@ -421,7 +415,7 @@ export function drawer() {
       this.elements.bound.strokeColor.alpha = this.state.selected ? 1 : 0.6;
       const { prevPosition } = this.data;
       if (!prevPosition) return;
-      const coordinate = mapOverlayLayer.globalToLocal(event.point);
+      const coordinate = layers.mapOverlayLayer.globalToLocal(event.point);
 
       // if the object was clicked, not dragged
       if (!this.data.wasMoved) {
@@ -450,110 +444,9 @@ export function drawer() {
     });
   }
 
-  mapLayer.activate();
+  layers.mapLayer.activate();
 
   // Todo: make state use a Listener paradigm rather than triggering method calls
-  const toolState = {
-    activeTool: null,
-    toolMap: {},
-
-    selected: {},
-    isSomethingSelected() {
-      return Object.keys(this.selected).length > 0;
-    },
-    isCanvasFocused: false,
-    toolIsActive: false,
-    isDown: false,
-    toolMapValue(definition, tool, modifiers) {
-      return {
-        type: definition.type,
-        definition,
-        tool,
-        modifiers,
-      };
-    },
-    defaultToolMapValue(toolType) {
-      const def = toolCategoryDefinition[toolType];
-      return this.toolMapValue(def, def.defaultTool, def.defaultModifiers);
-    },
-    switchToolType(toolType) {
-      if (!this.toolMap.hasOwnProperty(toolType)) {
-        this.switchTool(this.defaultToolMapValue(toolType), true);
-      } else {
-        this.switchTool(this.toolMap[toolType], true);
-      }
-    },
-    switchTool(toolData, isToolTypeSwitch) {
-      const prevTool = this.activeTool;
-      this.activeTool = toolData;
-      this.toolMap[toolData.type] = toolData;
-      if (prevTool) {
-        prevTool.definition.updateTool(prevTool, toolData, isToolTypeSwitch);
-      } else if (toolData) {
-        toolData.definition.updateTool(prevTool, toolData, isToolTypeSwitch);
-      }
-    },
-    deleteSelection() {
-      Object.keys(this.selected).forEach((objectId) => {
-        const object = this.selected[objectId];
-        object.onDelete();
-      });
-      this.deselectAll();
-    },
-    selectObject(object) {
-      this.deselectAll();
-      this.selected[object.data.id] = object;
-      object.onSelect(true);
-    },
-    deselectObject(object) {
-      delete this.selected[object.data.id];
-      object.onSelect(false);
-    },
-    deselectAll() {
-      Object.keys(this.selected).forEach((objectId) => {
-        const object = this.selected[objectId];
-        object.onSelect(false);
-      });
-      this.selected = {};
-    },
-    onDown(event) {
-      // deactivate the tool when something is selected or dragging an object
-      this.isDown = true;
-
-      // if we didn't click on one of the selected objects, deselect them
-      let clickedOnSelected = false;
-      Object.keys(this.selected).forEach((objectId) => {
-        const object = this.selected[objectId];
-        if (object.contains(mapOverlayLayer.globalToLocal(event.point))) {
-          clickedOnSelected = true;
-        }
-      });
-      if (!clickedOnSelected) {
-        this.deselectAll();
-      }
-    },
-    onUp(event) {
-      this.isDown = false;
-
-      const isActive = this.isCanvasFocused && !this.isSomethingSelected();
-      if (this.toolIsActive != isActive) {
-        this.toolIsActive = isActive;
-        if (this.activeTool) this.activeTool.definition.enablePreview(isActive);
-      }
-    },
-    focusOnCanvas(isFocused) {
-      this.isCanvasFocused = isFocused;
-      if (!this.isDown) {
-        const isActive = this.isCanvasFocused && !this.isSomethingSelected();
-        if (this.toolIsActive != isActive) {
-          this.toolIsActive = isActive;
-          if (this.activeTool) {
-            this.activeTool.definition.enablePreview(isActive);
-          }
-        }
-      }
-    },
-  };
 
   // ===============================================
   // GLOBAL FUNCTIONS
@@ -571,37 +464,37 @@ export function drawer() {
   let prevViewMatrix = paper.view.matrix.clone();
 
   const inverted = paper.view.matrix.inverted();
-  fixedLayer.activate();
+  layers.fixedLayer.activate();
 
-  //  cloudLayer.matrix = paper.view.matrix.inverted();
-  //  cloudLayer.scale(2, paper.view.projectToView(new paper.Point(0, 0)));
-  //  cloudLayer.bounds.topLeft = paper.view.projectToView(new paper.Point(0, 0));
+  //  layers.cloudLayer.matrix = paper.view.matrix.inverted();
+  //  layers.cloudLayer.scale(2, paper.view.projectToView(new paper.Point(0, 0)));
+  //  layers.cloudLayer.bounds.topLeft = paper.view.projectToView(new paper.Point(0, 0));
   function onFrame() {
     if (!paper.view.matrix.equals(prevViewMatrix)) {
       const inverted = paper.view.matrix.inverted();
       backgroundLayer.matrix = inverted;
 
-      fixedLayer.matrix = inverted;
+      layers.fixedLayer.matrix = inverted;
       modalLayer.matrix = inverted;
       prevViewMatrix = paper.view.matrix.clone();
 
       //      // clouds shift w/ parallax while scrolling
-      //      cloudLayer.matrix = inverted;
-      //      cloudLayer.scale(2, paper.view.projectToView(new paper.Point(0, 0)));
-      //      cloudLayer.bounds.topLeft = paper.view.projectToView(new paper.Point(0, 0));
+      //      layers.cloudLayer.matrix = inverted;
+      //      layers.cloudLayer.scale(2, paper.view.projectToView(new paper.Point(0, 0)));
+      //      layers.cloudLayer.bounds.topLeft = paper.view.projectToView(new paper.Point(0, 0));
     }
-    // fixedLayer.pivot = new paper.Point(0, 0);
-    // fixedLayer.position = paper.view.viewSize.topLeft;
+    // layers.fixedLayer.pivot = new paper.Point(0, 0);
+    // layers.fixedLayer.position = paper.view.viewSize.topLeft;
     // var inverseZoom = 1 / paper.view.zoom;
 
-    // fixedLayer.scaling = new paper.Point(inverseZoom, inverseZoom);
+    // layers.fixedLayer.scaling = new paper.Point(inverseZoom, inverseZoom);
   }
-  mapLayer.activate();
+  layers.mapLayer.activate();
 
   // ===============================================
   // BACKGROUND
 
-  //  cloudLayer.activate();
+  //  layers.cloudLayer.activate();
   //  for (var i = 0; i < 20; i ++) {
   //    var cloud = new paper.Raster('static/img/cloud1.png');
   //    cloud.position = new paper.Point((i % 2 + 2) * 120, (i % 2 + 3) * 120);
@@ -663,7 +556,7 @@ export function drawer() {
       ),
       new paper.Point(0, paper.view.size.height * paper.view.scaling.y),
     ];
-    mapLayer.activate();
+    layers.mapLayer.activate();
   }
 
   function jumpTween(item) {
@@ -673,7 +566,7 @@ export function drawer() {
   // ===============================================
   // MAIN UI
   window.addEventListener('beforeunload', (e) => {
-    if (actionsSinceSave == 0) {
+    if (state.actionsSinceSave === 0) {
       return undefined;
     }
 
@@ -707,7 +600,7 @@ export function drawer() {
   function autosaveMap() {
     if (localStorage) {
       localStorage.setItem('autosave', encodeMap());
-      actionsSinceSave = 0;
+      state.actionsSinceSave = 0;
       return true;
     }
     console.log(
@@ -743,13 +636,15 @@ export function drawer() {
 
     const saveMargins = new paper.Point(10, 10);
 
-    uiLayer.activate();
-    const mapRaster = mapLayer.rasterize();
-    const mapPositionDelta = mapLayer.globalToLocal(mapLayer.bounds.topLeft);
+    layers.uiLayer.activate();
+    const mapRaster = layers.mapLayer.rasterize();
+    const mapPositionDelta = layers.mapLayer.globalToLocal(
+      layers.mapLayer.bounds.topLeft,
+    );
 
-    const iconsRaster = mapIconLayer.rasterize();
-    const iconsPositionDelta = mapIconLayer.globalToLocal(
-      mapIconLayer.bounds.topLeft,
+    const iconsRaster = layers.mapIconLayer.rasterize();
+    const iconsPositionDelta = layers.mapIconLayer.globalToLocal(
+      layers.mapIconLayer.bounds.topLeft,
     );
 
     const gridRaster = getGridRaster();
@@ -757,8 +652,8 @@ export function drawer() {
     const gridClone = gridRaster.clone();
 
     const mapBounds = gridRaster.bounds.clone();
-    mapBounds.size += saveMargins;
-    mapBounds.point -= saveMargins / 2;
+    mapBounds.size.add(saveMargins);
+    mapBounds.point.subtract(saveMargins.divide(2));
     const mapBoundsClippingMask = new paper.Path.Rectangle(mapBounds);
 
     const background = mapBoundsClippingMask.clone();
@@ -767,7 +662,7 @@ export function drawer() {
     mapBoundsClippingMask.clipMask = true;
 
     const text = new paper.PointText(
-      mapBounds.bottomRight - new paper.Point(2, 2),
+      mapBounds.bottomRight.subtract(new paper.Point(2, 2)),
     );
     text.justification = 'right';
     text.content = 'made at eugeneration.github.io/HappyIslandDesigner';
@@ -790,10 +685,10 @@ export function drawer() {
     ]);
 
     // the raster doesn't scale for some reason, so manually scale it;
-    mapRaster.scaling /= mapLayer.scaling;
+    mapRaster.scaling.divide(layers.mapLayer.scaling);
     mapRaster.bounds.topLeft = mapPositionDelta;
 
-    iconsRaster.scaling /= mapLayer.scaling;
+    iconsRaster.scaling.divide(layers.mapLayer.scaling);
     iconsRaster.bounds.topLeft = iconsPositionDelta;
 
     const combinedImage = group.rasterize(708.5);
@@ -827,14 +722,14 @@ export function drawer() {
   }
 
   function loadMapFromFile() {
-    readFile = function (e) {
+    const readFile = function (e) {
       const file = e.target.files[0];
       if (!file) {
         return;
       }
       const reader = new FileReader();
       reader.onload = function (e) {
-        const dataURL = e.target.result;
+        const dataURL = e.target!.result as string;
 
         const image = new Image();
         image.src = dataURL;
@@ -898,86 +793,12 @@ export function drawer() {
   // highlightedColor: string
   // selectedColor: string
 
-  function createButton(item, buttonSize, onClick, options?: any) {
-    const highlightedColor =
-      !options || options.highlightedColor == null
-        ? colors.sand.color
-        : options.highlightedColor;
-    const selectedColor =
-      !options || options.selectedColor == null
-        ? colors.npc.color
-        : options.selectedColor;
-
-    const group = new paper.Group();
-
-    const button = new paper.Path.Circle(0, 0, buttonSize);
-
-    group.applyMatrix = false;
-    group.addChildren([button, item]);
-
-    function updateColor() {
-      button.fillColor =
-        group.data.selected || group.data.pressed
-          ? selectedColor
-          : highlightedColor;
-      button.fillColor.alpha = group.data.selected
-        ? 1
-        : group.data.pressed
-        ? 0.5
-        : group.data.hovered
-        ? 1
-        : 0.0001;
-    }
-    updateColor();
-
-    group.data = {
-      selected: false,
-      hovered: false,
-      pressed: false,
-      disabled: false,
-      select(isSelected) {
-        group.data.selected = isSelected;
-        updateColor();
-      },
-      hover(isHover) {
-        group.data.hovered = isHover;
-        updateColor();
-      },
-      press(isPressed) {
-        group.data.pressed = isPressed;
-        updateColor();
-      },
-      disable(isDisabled) {
-        group.data.disabled = isDisabled;
-        item.opacity = isDisabled ? 0.5 : 1;
-        if (isDisabled) group.data.hover(false);
-      },
-    };
-    group.onMouseEnter = function (event) {
-      if (group.data.disabled) return;
-      group.data.hover(true);
-    };
-    group.onMouseLeave = function (event) {
-      if (group.data.disabled) return;
-      group.data.press(false);
-      group.data.hover(false);
-    };
-    group.onMouseDown = function (event) {
-      if (group.data.disabled) return;
-      group.data.press(true);
-    };
-    group.onMouseUp = function (event) {
-      if (group.data.disabled) return;
-      if (group.data.pressed) onClick(event, group);
-      group.data.press(false);
-    };
-    return group;
-  }
-
   // ===============================================
   // TOOLS
+  createLeftMenu();
 
-  fixedLayer.activate();
+  layers.fixedLayer.activate();
+  init();
 
   // var menuButton = new paper.Path();
   // menuButton.strokeColor = colors.selected.color;
@@ -988,383 +809,6 @@ export function drawer() {
   //  new paper.Point(-20, 0),
   //  new paper.Point(0, 0),
   // ];
-
-  function renderModal(name, width, height, onDismiss) {
-    const topLeft = new paper.Point(0, 0); // + paper.view.bounds.topLeft;
-    const center = new paper.Point(
-      (paper.view.bounds.width * paper.view.scaling.x) / 2,
-      (paper.view.bounds.height * paper.view.scaling.y) / 2,
-    ); // + paper.view.bounds.topLeft * 2;
-    const bottomRight = new paper.Point(
-      paper.view.bounds.width * paper.view.scaling.x,
-      paper.view.bounds.height * paper.view.scaling.y,
-    ); // + paper.view.bounds.topLeft * 2;
-
-    modalLayer.activate();
-
-    const group = new paper.Group();
-
-    const darkFill = new paper.Path.Rectangle(
-      new paper.Rectangle(topLeft, bottomRight),
-    );
-    darkFill.fillColor = colors.offBlack.color;
-    darkFill.fillColor.alpha = 0.3;
-    darkFill.onMouseUp = onDismiss;
-
-    const modal = new paper.Path.Rectangle(
-      new paper.Rectangle(
-        center.x - width / 2,
-        center.y - height / 2,
-        width,
-        height,
-      ),
-      new paper.Size(60, 60),
-    );
-    modal.fillColor = colors.paper.color;
-    modal.onMouseEnter = function (event) {
-      group.data.text.content = name;
-    };
-
-    const modalContents = new paper.Group();
-    modalContents.applyMatrix = false;
-    modalContents.pivot = new paper.Point(0, 0);
-    modalContents.position = modal.bounds.topLeft + new paper.Point(40, 120);
-    modalContents.data = {
-      addElement() {},
-    };
-
-    group.data = {
-      width: modal.bounds.width - 40 * 2,
-      height: modal.bounds.height - 120 - 40,
-      contents: modalContents,
-    };
-
-    emitter.on('resize', () => {
-      const topLeft = new paper.Point(0, 0); // + paper.view.bounds.topLeft;
-      const center = new paper.Point(
-        (paper.view.bounds.width * paper.view.scaling.x) / 2,
-        (paper.view.bounds.height * paper.view.scaling.y) / 2,
-      ); // + paper.view.bounds.topLeft * 2;
-      const bottomRight = new paper.Point(
-        paper.view.bounds.width * paper.view.scaling.x,
-        paper.view.bounds.height * paper.view.scaling.y,
-      ); // + paper.view.bounds.topLeft * 2;
-
-      // var topLeft = paper.view.viewToProject(paper.view.projectToView(new paper.Point(0, 0)));// + paper.view.bounds.topLeft;
-      // var center = paper.view.viewToProject(paper.view.projectToView(new paper.Point(paper.view.bounds.width / 2, paper.view.bounds.height / 2)));// + paper.view.bounds.topLeft * 2;
-      // var bottomRight = paper.view.viewToProject(paper.view.projectToView(new paper.Point(paper.view.bounds.width, paper.view.bounds.height)));// + paper.view.bounds.topLeft * 2;
-
-      darkFill.bounds = new paper.Rectangle(topLeft, bottomRight);
-      modal.position = center;
-      modalContents.position = modal.bounds.topLeft + new paper.Point(40, 135);
-    });
-
-    const text = new paper.PointText(
-      new paper.Point(group.data.width / 2, -50),
-    );
-    text.justification = 'center';
-    (text.content = name), (text.fontSize = 20);
-    text.fontFamily = 'TTNorms, sans-serif';
-    text.fillColor = colors.text.color;
-    modalContents.addChild(text);
-
-    const statusBar = new paper.Raster('static/img/ui-phonestatus.png');
-    statusBar.scaling = 0.35;
-    statusBar.position = new paper.Point(group.data.width / 2 - 10, -93);
-    modalContents.addChild(statusBar);
-
-    const time = new paper.PointText(
-      new paper.Point(group.data.width / 2, -90),
-    );
-    time.justification = 'center';
-    time.fontSize = 12;
-    time.fontFamily = 'TTNorms, sans-serif';
-    time.fillColor = colors.lightText.color;
-    time.content = new Date().toLocaleTimeString([], {
-      hour: '2-digit',
-      minute: '2-digit',
-    });
-    setInterval(() => {
-      time.content = new Date().toLocaleTimeString([], {
-        hour: '2-digit',
-        minute: '2-digit',
-      });
-    }, 10000);
-    modalContents.addChild(time);
-
-    group.addChildren([darkFill, modal, modalContents]);
-
-    group.data.text = text;
-
-    return group;
-  }
-
-  let helpMenu;
-  function showHelpMenu(isShown) {
-    if (helpMenu == null) {
-      helpMenu = renderModal('Hotkeys', 340, 560, () => {
-        showHelpMenu(false);
-      });
-      helpMenu.onMouseUp = function () {
-        showHelpMenu(false);
-      };
-
-      const helpText = new paper.PointText(new paper.Point(80, -10));
-      helpText.justification = 'right';
-      helpText.fontSize = 16;
-      helpText.fontFamily = 'TTNorms, sans-serif';
-      helpText.fillColor = colors.oceanText.color;
-      helpText.content =
-        'space+drag\n' +
-        'alt+scroll\n' +
-        '\\\n' +
-        'shift+drag\n' +
-        '[ ]\n' +
-        'p\n' +
-        'alt+click\n' +
-        'delete\n' +
-        'ctrl + z\n' +
-        'ctrl + y\n' +
-        '\n' +
-        'v\n' +
-        'b\n' +
-        'n\n' +
-        'm\n' +
-        '\n' +
-        'ctrl + s\n' +
-        'ctrl + o\n' +
-        'esc\n' +
-        '?\n' +
-        '';
-
-      const helpText2 = new paper.PointText(new paper.Point(100, -10));
-      helpText2.justification = 'left';
-      helpText2.fontSize = 16;
-      helpText2.fontFamily = 'TTNorms, sans-serif';
-      helpText2.fillColor = colors.text.color;
-      helpText2.content =
-        'pan\n' +
-        'zoom\n' +
-        'toggle grid\n' +
-        'draw line\n' +
-        'adjust brush size\n' +
-        'square/circle brush\n' +
-        'color pick\n' +
-        'delete selection\n' +
-        'undo\n' +
-        'redo\n' +
-        '\n' +
-        'terrain tool \n' +
-        'path tool\n' +
-        'building tool\n' +
-        'amenities tool\n' +
-        '\n' +
-        'save\n' +
-        'open map file\n' +
-        'main menu\n' +
-        'hotkeys\n' +
-        '';
-
-      const helpTextRaster = helpText.rasterize();
-      const helpText2Raster = helpText2.rasterize();
-      helpText.remove();
-      helpText2.remove();
-
-      const versionCode = new paper.PointText(
-        helpMenu.data.width / 2,
-        helpMenu.data.height,
-      );
-      versionCode.justification = 'center';
-      versionCode.fontSize = 12;
-      versionCode.fontFamily = 'TTNorms, sans-serif';
-      versionCode.fillColor = colors.lightText.color;
-      versionCode.content = 'v0.2.1';
-
-      helpMenu.data.contents.addChildren([
-        helpTextRaster,
-        helpText2Raster,
-        versionCode,
-      ]);
-
-      helpMenu.opacity = 0;
-    }
-    helpMenu.tweenTo({ opacity: isShown ? 1 : 0 }, 200);
-    helpMenu.locked = !isShown;
-  }
-
-  let mainMenu;
-
-  function showMainMenu(isShown) {
-    if (mainMenu == null) {
-      if (!isShown) return;
-      mainMenu = renderModal('Main Menu', 260, 370, () => {
-        showMainMenu(false);
-      });
-
-      const hitSizeHalf = new paper.Point(35, 35);
-      const hitSize = new paper.Size(70, 70);
-      function createMenuButton(
-        name,
-        img,
-        index,
-        onMouseDown: () => void,
-        onMouseEnter?: () => void,
-      ) {
-        const buttonGroup = new paper.Group();
-
-        const button = new paper.Raster(img);
-        button.scaling = new paper.Point(0.4, 0.4);
-        button.locked = true;
-
-        const hitTarget = new paper.Path.Rectangle(
-          button.position - hitSizeHalf,
-          hitSize,
-        );
-        hitTarget.fillColor = colors.invisible.color;
-
-        buttonGroup.applyMatrix = false;
-        buttonGroup.addChildren([hitTarget, button]);
-        buttonGroup.position = new paper.Point(20 + index * 70, 0);
-
-        buttonGroup.onMouseDown = function (event) {
-          onMouseDown();
-        };
-
-        buttonGroup.onMouseEnter = function (event) {
-          mainMenu.data.text.content = name;
-
-          button.position = new paper.Point(0, 0);
-          button.animate([
-            {
-              properties: {
-                position: { y: '-5' },
-                scale: 1.1,
-              },
-              settings: {
-                duration: 60,
-                easing: 'linear',
-              },
-            },
-            {
-              properties: {
-                position: { y: '+7' },
-              },
-              settings: {
-                duration: 60,
-                easing: 'linear',
-              },
-            },
-            {
-              properties: {
-                position: { y: '-2' },
-              },
-              settings: {
-                duration: 120,
-                easing: 'linear',
-              },
-            },
-          ]);
-        };
-        buttonGroup.onMouseLeave = function (event) {
-          button.animate({
-            properties: {
-              scale: 1,
-            },
-            settings: {
-              duration: 60,
-              easing: 'linear',
-            },
-          });
-        };
-
-        return buttonGroup;
-      }
-
-      const saveButton = createMenuButton(
-        'Save as Image',
-        'static/img/menu-save.png',
-        0,
-        () => {
-          saveMapToFile();
-        },
-      );
-      const loadButton = createMenuButton(
-        'Load Map',
-        'static/img/menu-open.png',
-        1,
-        () => {
-          loadMapFromFile();
-        },
-      );
-      const newButton = createMenuButton(
-        'New Map',
-        'static/img/menu-new.png',
-        2,
-        () => {
-          const r = confirm(
-            'Clear your map? You will lose all unsaved changes.',
-          );
-          if (r == true) {
-            loadTemplate();
-          } else {
-          }
-        },
-      );
-
-      const twitterButton = createMenuButton(
-        'Twitter',
-        'static/img/menu-twitt.png',
-        0,
-        () => {
-          window.open('https://twitter.com/island_designer', '_blank');
-        },
-      );
-      twitterButton.position = new paper.Point(0, 210);
-
-      mainMenu.data.contents.addChildren([
-        saveButton,
-        loadButton,
-        newButton,
-        twitterButton,
-      ]);
-      mainMenu.opacity = 0;
-    }
-    mainMenu.tweenTo({ opacity: isShown ? 1 : 0 }, 200);
-    mainMenu.locked = !isShown;
-  }
-
-  const leftToolMenu = new paper.Group();
-  leftToolMenu.applyMatrix = false;
-  leftToolMenu.position = [30, 0];
-
-  const leftToolMenuBacking = new paper.Path();
-  leftToolMenuBacking.strokeColor = colors.paper.color;
-  leftToolMenuBacking.strokeWidth = 120;
-  leftToolMenuBacking.strokeCap = 'round';
-  leftToolMenuBacking.segments = [
-    new paper.Point(-30, -0),
-    new paper.Point(-30, 480),
-  ];
-  leftToolMenu.addChild(leftToolMenuBacking);
-
-  const leftToolMenuPosition = new paper.Point(0, 100);
-  const leftToolMenuIconHeight = 50;
-
-  function addToLeftToolMenu(icon?: any) {
-    if (icon == null) {
-      //      // create spacer
-      //      icon = new paper.Path.Rectangle(0, 0, 40, 2);
-      //      icon.fillColor = colors.lightText.color;
-      //      icon.position = leftToolMenuPosition - new paper.Point(0, leftToolMenuIconHeight / 4);
-      //      leftToolMenu.addChild(icon);
-      leftToolMenuPosition.y += leftToolMenuIconHeight / 2;
-      return;
-    }
-
-    icon.position = leftToolMenuPosition;
-    leftToolMenu.addChild(icon);
-    leftToolMenuPosition.y += leftToolMenuIconHeight;
-  }
 
   const redoButton = undoMenuButton('static/img/menu-redo.png', () => {
     redo();
@@ -1449,447 +893,6 @@ export function drawer() {
 
   // =======================================
   // BASE LEVEL TOOLS
-  const baseObjectCategoryDefinition = {
-    base: baseToolCategoryDefinition,
-    // type: 'tree', // filled in by base class
-    // icon: "amenities",
-    // tools: asyncTreeDefinition,
-    // menuOptions: {},
-    // yPos: 185
-    iconMenu: null,
-    layer: mapIconLayer,
-    defaultTool: null,
-    modifiers: {},
-    defaultModifiers: {},
-    onSelect(isSelected, isReselected) {
-      this.base.onSelect(this, isSelected, isReselected);
-    },
-    onMouseMove(event) {
-      this.base.onMouseMove(this, event);
-    },
-    onMouseDown(event) {
-      placeObject(event);
-      this.base.onMouseDown(this, event);
-    },
-    onMouseDrag(event) {
-      this.base.onMouseDrag(this, event);
-    },
-    onMouseUp(event) {
-      this.base.onMouseUp(this, event);
-    },
-    onKeyDown(event) {
-      this.base.onKeyDown(this, event);
-    },
-    enablePreview(isEnabled) {
-      this.base.enablePreview(this, isEnabled);
-      if (objectPreviewOutline) objectPreviewOutline.visible = isEnabled;
-      if (objectPreview) objectPreview.visible = isEnabled;
-    },
-    openMenu(isSelected) {
-      if (this.iconMenu == null) {
-        this.tools.getAsyncValue((definitions) => {
-          fixedLayer.activate();
-          const categoryDefinition = this;
-          this.iconMenu = createMenu(
-            objectMap(definitions, (def, name) => {
-              if (def.legacy || def.legacyCategory) return null;
-              const icon = createObjectIcon(def, getObjectData(def));
-              icon.scaling = def.menuScaling;
-              return createButton(icon, 20, (event, button) => {
-                toolState.switchTool(
-                  toolState.toolMapValue(categoryDefinition, def, {}),
-                );
-              });
-            }),
-            this.menuOptions,
-          );
-          this.iconMenu.data.setPointer(this.yPos);
-          this.iconMenu.pivot = new paper.Point(0, 0);
-          this.iconMenu.position = new paper.Point(100, 45);
-          // this is a little messy
-          if (toolState.activeTool && toolState.activeTool.tool) {
-            this.iconMenu.data.update(toolState.activeTool.tool.type);
-          }
-          this.iconMenu.visible = isSelected;
-        });
-      } else {
-        this.iconMenu.visible = isSelected;
-      }
-    },
-  };
-
-  amenitiesDef.load();
-  structureDef.load();
-  constructionDef.load();
-  treeDef.load();
-  flowerDef.load();
-
-  const toolCategoryDefinition = {
-    //    pointer: {
-    //      base: baseToolCategoryDefinition,
-    //      type: 'pointer',
-    //      layer: mapIconLayer,
-    //      icon: "pointer",
-    //      tools: {},
-    //      defaultTool: null,
-    //      modifiers: {},
-    //      defaultModifiers: {
-    //      },
-    //      onSelect: function(isSelected) {
-    //        this.base.onSelect(this, isSelected);
-    //      },
-    //      onMouseMove: function(event) {
-    //        this.base.onMouseMove(this, event);
-    //      },
-    //      onMouseDown: function(event) {
-    //        this.base.onMouseDown(this, event);
-    //      },
-    //      onMouseDrag: function(event) {
-    //        this.base.onMouseDrag(this, event);
-    //      },
-    //      onMouseUp: function(event) {
-    //        this.base.onMouseUp(this, event);
-    //      },
-    //      onKeyDown: function(event) {
-    //        this.base.onKeyDown(this, event);
-    //      },
-    //      enablePreview: function(isEnabled) {
-    //        this.base.enablePreview(this, isEnabled);
-    //      },
-    //    },
-    terrain: {
-      base: baseToolCategoryDefinition,
-      type: 'terrain',
-      layer: mapLayer,
-      icon: 'color',
-      modifiers: {},
-      iconMenu: null,
-      defaultModifiers: {},
-      data: {
-        paintColorData: colors.level1,
-      },
-      onSelect(isSelected, isReselected) {
-        this.base.onSelect(this, isSelected, isReselected);
-      },
-      onMouseMove(event) {
-        this.base.onMouseMove(this, event);
-      },
-      onMouseDown(event) {
-        this.base.onMouseDown(this, event);
-        if (paper.Key.isDown('alt')) {
-          const rawCoordinate = mapOverlayLayer.globalToLocal(event.point);
-          updatePaintColor(getColorAtCoordinate(rawCoordinate));
-        }
-        startDraw(event);
-      },
-      onMouseDrag(event) {
-        this.base.onMouseDrag(this, event);
-        draw(event);
-      },
-      onMouseUp(event) {
-        this.base.onMouseUp(this, event);
-        endDraw(event);
-      },
-      onKeyDown(event) {
-        this.base.onKeyDown(this, event);
-      },
-      enablePreview(isEnabled) {
-        this.base.enablePreview(this, isEnabled);
-        brushOutline.visible = isEnabled;
-        brush.visible = isEnabled;
-      },
-      openMenu(isSelected) {
-        if (this.iconMenu == null) {
-          fixedLayer.activate();
-          updatePaintColor(this.data.paintColorData);
-          this.iconMenu = createMenu(
-            objectMap(layerDefinition, (definition, colorKey) => {
-              const colorData = colors[colorKey];
-              const paintCircle = new paper.Path.Circle(
-                new paper.Point(0, 0),
-                16,
-              );
-              paintCircle.fillColor = colorData.color;
-              paintCircle.locked = true;
-              return createButton(paintCircle, 20, (event, button) => {
-                updatePaintColor(colorData);
-                this.data.paintColorData = colorData;
-              });
-            }),
-            { spacing: 45, extraColumns: 1 },
-          );
-          this.iconMenu.data.setPointer(60);
-          this.iconMenu.pivot = new paper.Point(0, 0);
-          this.iconMenu.position = new paper.Point(100, 45);
-          // this is a little messy
-          this.iconMenu.data.update(this.data.paintColorData.key);
-        }
-        this.iconMenu.visible = isSelected;
-        const adjusterUI = showBrushSizeUI(isSelected);
-      },
-    },
-    path: {
-      base: baseToolCategoryDefinition,
-      type: 'path',
-      layer: mapLayer,
-      icon: 'path',
-      modifiers: {},
-      defaultModifiers: {},
-      iconMenu: null,
-      data: {
-        paintColorData: colors.pathDirt,
-      },
-      onSelect(isSelected, isReselected) {
-        this.base.onSelect(this, isSelected, isReselected);
-      },
-      onMouseMove(event) {
-        this.base.onMouseMove(this, event);
-      },
-      onMouseDown(event) {
-        this.base.onMouseDown(this, event);
-        if (paper.Key.isDown('alt')) {
-          const rawCoordinate = mapOverlayLayer.globalToLocal(event.point);
-          updatePaintColor(getColorAtCoordinate(rawCoordinate));
-        }
-        startDraw(event);
-      },
-      onMouseDrag(event) {
-        this.base.onMouseDrag(this, event);
-        draw(event);
-      },
-      onMouseUp(event) {
-        this.base.onMouseUp(this, event);
-        endDraw(event);
-      },
-      onKeyDown(event) {
-        this.base.onKeyDown(this, event);
-      },
-      enablePreview(isEnabled) {
-        this.base.enablePreview(this, isEnabled);
-        brushOutline.visible = isEnabled;
-        brush.visible = isEnabled;
-      },
-      openMenu(isSelected) {
-        if (this.iconMenu == null) {
-          fixedLayer.activate();
-          updatePaintColor(this.data.paintColorData);
-          const pathColorButtons = objectMap(
-            pathDefinition,
-            (definition, colorKey) => {
-              let buttonIcon;
-              const colorData = colors[colorKey];
-              if (colorKey == colors.pathEraser.key) {
-                buttonIcon = new paper.Group();
-                const eraserImg = new paper.Raster(
-                  `${imgPath + toolPrefix}eraser.png`,
-                );
-                eraserImg.scaling = new paper.Point(0.35, 0.35);
-                buttonIcon.addChildren([eraserImg]);
-              } else {
-                const paintCircle = new paper.Path.Circle(
-                  new paper.Point(0, 0),
-                  16,
-                );
-                paintCircle.fillColor = colorData.color;
-                paintCircle.locked = true;
-                buttonIcon = paintCircle;
-              }
-
-              return createButton(buttonIcon, 20, (event, button) => {
-                updatePaintColor(colorData);
-                this.data.paintColorData = colorData;
-              });
-            },
-          );
-          this.iconMenu = createMenu(pathColorButtons, {
-            spacing: 45,
-            extraColumns: 1,
-            extraRows: 1,
-          });
-          this.iconMenu.data.setPointer(110);
-          this.iconMenu.pivot = new paper.Point(0, 0);
-          this.iconMenu.position = new paper.Point(100, 45);
-          // this is a little messy
-          this.iconMenu.data.update(this.data.paintColorData.key);
-        }
-        this.iconMenu.visible = isSelected;
-        const adjusterUI = showBrushSizeUI(isSelected);
-      },
-    },
-    structures: Object.assign(Object.create(baseObjectCategoryDefinition), {
-      type: 'structures',
-      icon: 'structure',
-      tools: structureDef.asyncStructureDefinition,
-      menuOptions: { spacing: 50, perColumn: 9 },
-      yPos: 160,
-    }),
-    amenities: Object.assign(Object.create(baseObjectCategoryDefinition), {
-      type: 'amenities',
-      icon: 'amenities',
-      tools: amenitiesDef.asyncAmenitiesDefinition,
-      menuOptions: { spacing: 50, perColumn: 8 },
-      yPos: 208,
-    }),
-
-    construction: Object.assign(Object.create(baseObjectCategoryDefinition), {
-      type: 'construction',
-      icon: 'construction',
-      tools: constructionDef.asyncConstructionDefinition,
-      menuOptions: { spacing: 50, perColumn: 9 },
-      yPos: 260,
-    }),
-
-    tree: Object.assign(Object.create(baseObjectCategoryDefinition), {
-      type: 'tree',
-      icon: 'tree',
-      tools: treeDef.asyncTreeDefinition,
-      menuOptions: { spacing: 50, perColumn: 8 },
-      yPos: 310,
-    }),
-    flower: Object.assign(Object.create(baseObjectCategoryDefinition), {
-      type: 'flower',
-      icon: 'flower',
-      tools: flowerDef.asyncFlowerDefinition,
-      menuOptions: { spacing: 50, perColumn: 9 },
-      yPos: 360,
-    }),
-
-    //  shovel: {
-
-    // },
-    //  sprite: {
-    //    type: 'sprite',
-    //    targetLayers: [mapIconLayer],
-    //  },
-  };
-  // add additional sub functions to all definitions
-  Object.keys(toolCategoryDefinition).forEach((toolType) => {
-    const def = toolCategoryDefinition[toolType];
-    def.updateTool = function (prevToolData, nextToolData, isToolTypeSwitch) {
-      def.base.updateTool(def, prevToolData, nextToolData, isToolTypeSwitch);
-    };
-  });
-  const baseToolCategoryDefinition = {
-    onSelect(subclass, isSelected, isReselected) {
-      subclass.icon.data.select(isSelected);
-
-      if (isReselected) this.toggleMenu(subclass);
-      else this.openMenu(subclass, isSelected);
-
-      if (!isSelected) subclass.enablePreview(isSelected);
-    },
-    onMouseMove(subclass, event) {
-      updateCoordinateLabel(event);
-    },
-    onMouseDown(subclass, event) {
-      updateCoordinateLabel(event);
-    },
-    onMouseDrag(subclass, event) {
-      updateCoordinateLabel(event);
-    },
-    onMouseUp(subclass, event) {
-      updateCoordinateLabel(event);
-    },
-    onKeyDown(subclass, event) {},
-    enablePreview(subclass, isEnabled) {},
-    toggleMenu(subclass) {
-      if (subclass.openMenu) {
-        subclass.openMenu(!(subclass.iconMenu && subclass.iconMenu.visible));
-      }
-    },
-    openMenu(subclass, isSelected) {
-      if (subclass.openMenu) {
-        subclass.openMenu(isSelected);
-      }
-    },
-    updateTool(subclass, prevToolData, nextToolData, isToolTypeSwitch) {
-      const sameToolType =
-        prevToolData &&
-        prevToolData.definition.type === nextToolData.definition.type;
-      if (!sameToolType) {
-        if (prevToolData) {
-          prevToolData.definition.onSelect(false);
-        }
-        nextToolData.definition.onSelect(true);
-      } else if (isToolTypeSwitch) {
-        // user pressed the tool menu button - toggle the menu visibility
-        prevToolData.definition.onSelect(true, true);
-      }
-      {
-        const prevTool =
-          prevToolData && prevToolData.tool ? prevToolData.tool.type : null;
-        const nextTool =
-          nextToolData && nextToolData.tool ? nextToolData.tool.type : null;
-        const sameTool = sameToolType && prevTool === nextTool;
-        if (!sameTool) {
-          if (prevToolData && prevToolData.tool && prevToolData.tool.onSelect) {
-            prevToolData.tool.onSelect(false);
-          }
-          if (nextToolData && nextToolData.tool && nextToolData.tool.onSelect) {
-            nextToolData.tool.onSelect(true);
-          }
-          // todo: decouple view from logic
-          if (
-            subclass.iconMenu &&
-            (nextToolData.type == 'structures' ||
-              nextToolData.type == 'amenities' ||
-              nextToolData.type == 'construction' ||
-              nextToolData.type == 'tree' ||
-              nextToolData.type == 'flower')
-          ) {
-            subclass.iconMenu.data.update(nextTool);
-            updateObjectPreview();
-          }
-        }
-      }
-    },
-  };
-
-  // function squircle (size){ // squircle=square+circle
-  //  var hsize = size / 2; // half size
-  //
-  //  var squircle = new paper.Path();
-  //
-  //  squircle.add(
-  //    new Segment(new paper.Point(0,0), new paper.Point(0,0), new paper.Point(0,hsize)),
-  //    new Segment(new paper.Point(0,size), new paper.Point(0,size), new paper.Point(hsize,size)),
-  //    new Segment(new paper.Point(size,size), new paper.Point(size,size), new paper.Point(size,hsize)),
-  //    new Segment(new paper.Point(size,0), new paper.Point(size,0), new paper.Point(hsize,0))
-  //  );
-  //  squircle.closed = true;
-  //
-  //  return squircle;
-  // }
-  // fixedLayer.activate();
-  // var box = squircle(100);
-  // box.fillColor = colors.npc;
-  // box.position = new paper.Point(300, 300);
-  // box.selected = true;
-  //
-  // var d = new paper.Path.Rectangle(300, 300, 10, 10);
-  // d.fillColor = colors.npc;
-
-  // var activeToolIndicator = new paper.Path.Rectangle(0, 100, 5, 40);
-  // var activeToolIndicator = new paper.Path.Circle(30, 120, 20);
-  // activeToolIndicator.fillColor = colors.npc;
-
-  Object.keys(toolCategoryDefinition).forEach((toolType) => {
-    const def = toolCategoryDefinition[toolType];
-    const tool = new paper.Raster(`${imgPath + toolPrefix + def.icon}.png`);
-
-    const button = createButton(tool, 20, () => {
-      toolState.switchToolType(toolType);
-    });
-    switch (def.icon) {
-      case 'color':
-        tool.position = new paper.Point(-8, 0);
-        break;
-    }
-    tool.scaling = new paper.Point(0.4, 0.4);
-
-    addToLeftToolMenu(button);
-    def.icon = button;
-  });
   addToLeftToolMenu(); // spacer
 
   const toolBtn = new paper.Raster(`${imgPath}menu-help.png`);
@@ -1901,9 +904,6 @@ export function drawer() {
   };
   addToLeftToolMenu(button);
 
-  // add gap
-  leftToolMenuPosition.y += 60;
-
   // var activeColor = new paper.Path.Circle([20, 20], 16);
   // activeColor.fillColor = paintColor;
   // addToLeftToolMenu(activeColor);
@@ -1912,35 +912,7 @@ export function drawer() {
   //  activeColor
   // }
 
-  var paintColor = colors.level1;
-  function updatePaintColor(colorData) {
-    paintColor = colorData;
-    brush.fillColor = colorData.color;
-    // activeColor.fillColor = paintColor;
-
-    // todo: separate viewfrom logic
-    if (
-      (toolState.activeTool &&
-        toolState.activeTool.type == toolCategoryDefinition.terrain.type) ||
-      toolState.activeTool.type == toolCategoryDefinition.path.type
-    ) {
-      if (toolState.activeTool.definition.iconMenu) {
-        let toolCategory;
-        if (layerDefinition[colorData.key]) {
-          toolCategory = toolCategoryDefinition.terrain.type;
-        } else if (pathDefinition[colorData.key]) {
-          toolCategory = toolCategoryDefinition.path.type;
-        }
-        if (toolState.activeTool.type != toolCategory) {
-          toolState.switchToolType(toolCategory);
-        }
-
-        toolState.activeTool.definition.iconMenu.data.update(colorData.key);
-      }
-    }
-  }
-
-  fixedLayer.activate();
+  layers.fixedLayer.activate();
 
   const toolsPosition = new paper.Point(40, 80);
 
@@ -2095,7 +1067,7 @@ export function drawer() {
     }
   }
 
-  // mapOverlayLayer.activate();
+  // layers.mapOverlayLayer.activate();
   // var tracemap = new paper.Raster('static/img/tracemap.png');
   // tracemap.locked = true;
   // tracemap.position = new paper.Point(55.85, 52.2);
@@ -2193,7 +1165,7 @@ export function drawer() {
   }
 
   function decodeMap(json) {
-    mapLayer.activate();
+    layers.mapLayer.activate();
     const { version } = json;
     return {
       version: json.version,
@@ -2285,7 +1257,7 @@ export function drawer() {
   }
 
   function placeObject(event) {
-    const coordinate = mapOverlayLayer.globalToLocal(event.point);
+    const coordinate = layers.mapOverlayLayer.globalToLocal(event.point);
     if (toolState.activeTool && toolState.activeTool.tool) {
       const objectData = getObjectData(toolState.activeTool.tool);
       const command = objectCreateCommand(
@@ -2409,23 +1381,23 @@ export function drawer() {
       marginX = (viewWidth - width) / 2;
     }
 
-    mapLayer.position = new paper.Point(marginX, marginY);
-    mapLayer.scaling = new paper.Point(cellWidth, cellHeight);
+    layers.mapLayer.position = new paper.Point(marginX, marginY);
+    layers.mapLayer.scaling = new paper.Point(cellWidth, cellHeight);
 
-    mapOverlayLayer.position = new paper.Point(marginX, marginY);
-    mapOverlayLayer.scaling = new paper.Point(cellWidth, cellHeight);
+    layers.mapOverlayLayer.position = new paper.Point(marginX, marginY);
+    layers.mapOverlayLayer.scaling = new paper.Point(cellWidth, cellHeight);
 
-    mapIconLayer.position = new paper.Point(marginX, marginY);
-    mapIconLayer.scaling = new paper.Point(cellWidth, cellHeight);
+    layers.mapIconLayer.position = new paper.Point(marginX, marginY);
+    layers.mapIconLayer.scaling = new paper.Point(cellWidth, cellHeight);
   }
 
-  mapOverlayLayer.activate();
-  createGrid(mapOverlayLayer, mapLayer);
+  layers.mapOverlayLayer.activate();
+  createGrid();
 
   // ===============================================
   // COORDINATE LABEL
 
-  mapOverlayLayer.activate();
+  layers.mapOverlayLayer.activate();
   // var coordinateLabel = new paper.PointText(new paper.Point(0, 0));
   // coordinateLabel.fontSize = 3;
 
@@ -2442,7 +1414,7 @@ export function drawer() {
   // }
   //
 
-  mapLayer.activate();
+  layers.mapLayer.activate();
 
   if (!tryLoadAutosaveMap()) {
     loadTemplate();
@@ -2502,7 +1474,7 @@ export function drawer() {
   let prevGridCoordinate;
   let prevDrawCoordinate;
 
-  let diffCollection = {};
+  const diffCollection = {};
 
   // start/end: lattice Point
   // return: unioned Path/CompoundPath
@@ -2570,82 +1542,6 @@ export function drawer() {
     return linePath;
   }
 
-  // todo: merge this with the other preview code
-  let drawPreview;
-  function drawGridLinePreview(viewPosition) {
-    const rawCoordinate = new paper.Point(mapLayer.globalToLocal(viewPosition));
-    const coordinate = getBrushCenteredCoordinate(rawCoordinate);
-
-    mapLayer.activate();
-    if (drawPreview) {
-      drawPreview.remove();
-    }
-    if (startGridCoordinate == null) startDrawGrid(viewPosition);
-    drawPreview = drawLine(coordinate, startGridCoordinate);
-    if (drawPreview) {
-      drawPreview.locked = true;
-      drawPreview.opacity = 0.6;
-      drawPreview.fillColor = paintColor.color;
-    }
-  }
-
-  function stopGridLinePreview() {
-    if (drawPreview) drawPreview.remove();
-  }
-
-  function startDrawGrid(viewPosition) {
-    mapLayer.activate();
-    let coordinate = new paper.Point(mapLayer.globalToLocal(viewPosition));
-    coordinate = getBrushCenteredCoordinate(coordinate);
-    startGridCoordinate = coordinate;
-    prevGridCoordinate = coordinate;
-    drawGrid(viewPosition);
-  }
-
-  function drawGrid(viewPosition) {
-    mapLayer.activate();
-    const rawCoordinate = new paper.Point(mapLayer.globalToLocal(viewPosition));
-    const coordinate = getBrushCenteredCoordinate(rawCoordinate);
-
-    if (prevGridCoordinate == null) startDrawGrid(viewPosition);
-    const path = drawLine(coordinate, prevGridCoordinate);
-    if (path) {
-      const diff = getDiff(path, paintColor.key);
-
-      Object.keys(diff).forEach((colorKey) => {
-        const colorDiff = diff[colorKey];
-        if (!diffCollection.hasOwnProperty(colorKey)) {
-          diffCollection[colorKey] = { isAdd: colorDiff.isAdd, path: [] };
-        }
-        diffCollection[colorKey].path.push(colorDiff.path);
-        if (diffCollection[colorKey].isAdd != colorDiff.isAdd) {
-          console.error(`Simultaneous add and remove for ${colorKey}`);
-        }
-      });
-      applyDiff(true, diff);
-    }
-
-    prevGridCoordinate = coordinate;
-  }
-
-  function endDrawGrid(viewPosition) {
-    const mergedDiff = {};
-    prevGridCoordinate = null;
-    startGridCoordinate = null;
-    Object.keys(diffCollection).forEach((k) => {
-      mergedDiff[k] = {
-        isAdd: diffCollection[k].isAdd,
-        path: uniteCompoundPath(
-          new paper.CompoundPath({ children: diffCollection[k].path }),
-        ),
-      };
-    });
-    diffCollection = {};
-    if (Object.keys(mergedDiff).length > 0) {
-      addToHistory(drawCommand(mergedDiff));
-    }
-  }
-
   function uniteCompoundPath(compound) {
     let p = new paper.Path();
     compound.children.forEach((c) => {
@@ -2674,14 +1570,6 @@ export function drawer() {
       addDrawPoints(drawPoints, paintColor);
     }
   } */
-
-  function getDistanceFromWholeNumber(f) {
-    return Math.abs(f - Math.round(f));
-  }
-
-  function pointApproximates(p0, p1) {
-    return Math.abs(p0.x - p1.x) < 0.001 && Math.abs(p0.y - p1.y) < 0.001;
-  }
 
   function getDiff(path, colorKey) {
     if (!path.children && path.segments.length < 3) return {};
@@ -2802,7 +1690,7 @@ export function drawer() {
   }
 
   function addPath(isAdd, path, colorKey) {
-    mapLayer.activate();
+    layers.mapLayer.activate();
     if (!state.drawing.hasOwnProperty(colorKey)) {
       state.drawing[colorKey] = new paper.Path();
       state.drawing[colorKey].locked = true;
