+++ app/helpers/getDistanceFromWholeNumber.ts
@@ -1,0 +1,3 @@
+export function getDistanceFromWholeNumber(f) {
+  return Math.abs(f - Math.round(f));
+}
